updated go package from docs to swagger to make docs folder available for github pages moved swagger.yaml to docs/swagger.json and updated scripts to produce json integrated support for swagger path in albatross api
