+++ cmd/albatross/albatross.go
@@ -3,6 +3,8 @@ package main
 import (
 	"fmt"
 	"net/http"
+	"os"
+	"strconv"
 
 	"github.com/gorilla/mux"
 
@@ -11,7 +13,7 @@ import (
 	"github.com/gojekfarm/albatross/api/list"
 	"github.com/gojekfarm/albatross/api/uninstall"
 	"github.com/gojekfarm/albatross/api/upgrade"
-	_ "github.com/gojekfarm/albatross/docs"
+	_ "github.com/gojekfarm/albatross/swagger"
 	"github.com/gojekfarm/albatross/pkg/helmcli"
 	"github.com/gojekfarm/albatross/pkg/logger"
 
@@ -45,8 +47,18 @@ func startServer() {
 	router.Handle("/install", ContentTypeMiddle(installHandler)).Methods(http.MethodPut)
 	router.Handle("/upgrade", ContentTypeMiddle(upgradeHandler)).Methods(http.MethodPost)
 
+	serveDocumentation(router)
 	err := http.ListenAndServe(fmt.Sprintf(":%d", 8080), router)
 	if err != nil {
 		logger.Errorf("error starting server", err)
+	}
+}
+
+func serveDocumentation(r *mux.Router) {
+	docEnv := os.Getenv("DOCUMENTATION")
+	serveDoc, err := strconv.ParseBool(docEnv)
+	if err == nil && serveDoc {
+		fs := http.FileServer(http.Dir("./swaggerui"))
+		r.PathPrefix("/swaggerui/").Handler(http.StripPrefix("/swaggerui/", fs))
 	}
 }
