+++ scripts/swagger.sh
@@ -2,12 +2,12 @@
 
 create_doc() {
     GO111MODULE=off go get -v github.com/go-swagger/go-swagger/cmd/swagger
-    swagger generate spec -m -o swagger.yaml --exclude-deps
+    swagger generate spec -m -o docs/swagger.json --exclude-deps
 }
 
 validate_doc() {
     GO111MODULE=off go get -v github.com/go-swagger/go-swagger/cmd/swagger
-    swagger validate swagger.yaml
+    swagger validate docs/swagger.json
 }
 
 $*
